refactor(stringcalculator): Delimiter 관련 기능 분리
