+++ src/main/java/stringcalculator/StringCalculator.java
@@ -4,11 +4,6 @@ import java.util.Arrays;
 
 public class StringCalculator {
 
-    private static final StringCalculatorDelimiter BASE_DELIMITER = new StringCalculatorDelimiter(",|:");
-    private static final String CUSTOM_DELIMITER_HEADER = "//";
-    private static final String CUSTOM_DELIMITER_SUFFIX = "\n";
-    private static final String CUSTOM_DELIMITER_INPUT_PATTERN = "^//.*\n.+";
-
     private StringCalculator() {
         throw new IllegalStateException("Utility class");
     }
@@ -17,38 +12,15 @@ public class StringCalculator {
         if (StringHelper.isNullOrEmpty(input)) {
             return 0;
         }
-        var body = extractNumbersData(input);
-        var delimiter = extractDelimiterObject(input);
-        return sumDelimitedNumbers(body, delimiter);
+        var body = StringCalculatorDelimiter.removeCustomDelimiter(input);
+        var delimiter = StringCalculatorDelimiter.extractDelimiterObject(input);
+        return sumDelimitedNumbersData(body, delimiter);
     }
 
-    public static String extractNumbersData(String input) {
-        if (containsCustomDelimiter(input)) {
-            int realInputStart = input.indexOf(CUSTOM_DELIMITER_SUFFIX) + 1;
-            return input.substring(realInputStart);
-        }
-        return input;
-    }
-
-    public static StringCalculatorDelimiter extractDelimiterObject(String input) {
-        if (containsCustomDelimiter(input)) {
-            return BASE_DELIMITER.add(extractCustomDelimiter(input));
-        }
-        return BASE_DELIMITER;
-    }
-
-    public static String extractCustomDelimiter(String input) {
-        return input.substring(CUSTOM_DELIMITER_HEADER.length(), input.indexOf(CUSTOM_DELIMITER_SUFFIX));
-    }
-
-    public static int sumDelimitedNumbers(String realInput, StringCalculatorDelimiter delimiter) {
+    public static int sumDelimitedNumbersData(String realInput, StringCalculatorDelimiter delimiter) {
         return Arrays.stream(realInput.split(delimiter.value()))
                 .map(PositiveNumber::of)
                 .mapToInt(PositiveNumber::value)
                 .sum();
-    }
-
-    public static boolean containsCustomDelimiter(String input) {
-        return RegexCache.matches(CUSTOM_DELIMITER_INPUT_PATTERN, input);
     }
 }
