+++ src/test/java/stringcalculator/StringCalculatorTest.java
@@ -26,20 +26,6 @@ class StringCalculatorTest {
         assertThat(result).isZero();
     }
 
-    @DisplayName("커스텀 구분자는 '//'와 '\\n' 사이에 위치하며 커스텀 구분자를 추출한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {";", "&", "."})
-    void extract_custom_delimiter_between_double_slash_and_new_line(String delimiter) {
-        // given
-        String input = "//" + delimiter + "\n1" + delimiter + "2" + delimiter + "3";
-
-        // when
-        String customDelimiter = StringCalculator.extractCustomDelimiter(input);
-
-        // then
-        assertThat(customDelimiter).isEqualTo(delimiter);
-    }
-
     @DisplayName("쉼표(,)와 콜론(:) 및 커스텀 구분자 기준으로 숫자를 분리하여 덧셈을 수행한다.")
     @ParameterizedTest
     @ValueSource(strings = {"//;\n1;2;3", "1,2:3", "1:2:3", "//&\n1&2:3", "//.\n1.2.3"})
@@ -73,70 +59,6 @@ class StringCalculatorTest {
         assertThat(result).isEqualTo(Integer.parseInt(input));
     }
 
-    @DisplayName("입력한 문자열에서 숫자 데이터를 추출한다.")
-    @ParameterizedTest
-    @MethodSource("provideInputStringOfBody")
-    void extract_numbers_data_from_input_string(String input, String expected) {
-        // when
-        String result = StringCalculator.extractNumbersData(input);
-
-        // then
-        assertThat(result).isEqualTo(expected);
-    }
-
-    static Stream<Arguments> provideInputStringOfBody() {
-        return Stream.of(
-                Arguments.of("1,2,3", "1,2,3"),
-                Arguments.of("//;\n1;2;3", "1;2;3"),
-                Arguments.of("1:2:3", "1:2:3"),
-                Arguments.of("//&\n1&2:3", "1&2:3"),
-                Arguments.of("//.\n1.2.3", "1.2.3")
-        );
-    }
-
-    @DisplayName("입력한 문자열에서 커스텀 구분자가 포함되어 있는지 확인한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"//;\n1;2;3", "//&\n1&2:3", "//.\n1.2.3"})
-    void contains_custom_delimiter(String input) {
-        // when
-        boolean result = StringCalculator.containsCustomDelimiter(input);
-
-        // then
-        assertThat(result).isTrue();
-    }
-
-    @DisplayName("입력한 문자열에서 커스텀 구분자가 포함되어 있지 않은 경우, false를 반환한다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"1,2,3", "1:2:3", "1,2:3"})
-    void does_not_contain_custom_delimiter(String input) {
-        // when
-        boolean result = StringCalculator.containsCustomDelimiter(input);
-
-        // then
-        assertThat(result).isFalse();
-    }
-
-    @DisplayName("입력한 문자열에서 구분자 객체를 추출한다.")
-    @ParameterizedTest
-    @MethodSource("provideInputStringOfDefaultDelimiter")
-    void extract_delimiter_object_from_input_string(String input, String expected) {
-        // when
-        StringCalculatorDelimiter result = StringCalculator.extractDelimiterObject(input);
-
-        // then
-        assertThat(result.value()).isEqualTo(expected);
-    }
-
-    static Stream<Arguments> provideInputStringOfDefaultDelimiter() {
-        return Stream.of(
-                Arguments.of("1,2,3", ",|:"),
-                Arguments.of("//;\n1;2;3", ",|:|[;]"),
-                Arguments.of("1:2:3", ",|:"),
-                Arguments.of("//&\n1&2:3", ",|:|[&]"),
-                Arguments.of("//.\n1.2.3", ",|:|[.]")
-        );
-    }
-
     @DisplayName("숫자로 이루어진 문자열을 구분자로 나누어 합산한다.")
     @ParameterizedTest
     @MethodSource("provideInputStringOfSumNumbers")
@@ -145,7 +67,7 @@ class StringCalculatorTest {
         StringCalculatorDelimiter delimiter = new StringCalculatorDelimiter(delimiterStr);
 
         // when
-        int result = StringCalculator.sumDelimitedNumbers(input, delimiter);
+        int result = StringCalculator.sumDelimitedNumbersData(input, delimiter);
 
         // then
         assertThat(result).isEqualTo(expected);
